Refactor treasury page to v2
